+++ tests/dummy/app/locales/en/translations.js
@@ -42,6 +42,9 @@ Ember.$.extend(true, translations, {
           'commentsCount': {
             'caption': 'Comments count'
           },
+          'comments': {
+            'caption': 'Comments'
+          },
         },
         'SuggestionE': {
           'address': {
@@ -407,6 +410,10 @@ Ember.$.extend(true, translations, {
             'selected-rows': {
               'caption': 'Selected rows',
               'title': ''
+            },
+            'object-list-view-resize': {
+              'caption': 'Columns markup',
+              'title': ''
             }
           },
           'flexberry-simpledatetime': {
@@ -756,6 +763,11 @@ Ember.$.extend(true, translations, {
         },
         'selected-rows': {
           'caption': 'Flexberry-objectlistview. Setected rows'
+        },
+        'object-list-view-resize': {
+          'caption': 'Flexberry-objectlistview. Columns markup',
+          'button-сaption': 'Add',
+          'title': ''
         }
       },
       'flexberry-simpleolv': {