+++ tests/dummy/app/locales/ru/translations.js
@@ -42,6 +42,9 @@ Ember.$.extend(true, translations, {
           'commentsCount': {
             'caption': 'Количество комментариев'
           },
+          'comments': {
+            'caption': 'Комментарии'
+          },
         },
         'SuggestionE': {
           'address': {
@@ -424,6 +427,10 @@ Ember.$.extend(true, translations, {
             'selected-rows': {
               'caption': 'Выбранные строки',
               'title': ''
+            },
+            'object-list-view-resize': {
+              'caption': 'Разметка столбцов',
+              'title': ''
             }
           },
           'flexberry-simpledatetime': {
@@ -774,6 +781,11 @@ Ember.$.extend(true, translations, {
         },
         'selected-rows': {
           'caption': 'Flexberry-objectlistview. Выбранные строки'
+        },
+        'object-list-view-resize': {
+          'caption': 'Flexberry-objectlistview. Разметка столбцов',
+          'button-сaption': 'Добавить',
+          'title': ''
         }
       },
       'flexberry-simpleolv': {