+++ tests/dummy/app/routes/components-examples/flexberry-objectlistview/object-list-view-resize.js
@@ -1,0 +1,14 @@
+import ListFormRoute from 'ember-flexberry/routes/list-form';
+
+export default ListFormRoute.extend({
+
+  modelProjection: 'SuggestionL',
+
+  developerUserSettings: { ObjectListViewResize: { } },
+
+  modelName: 'ember-flexberry-dummy-suggestion',
+
+  model: function () {
+    return this.store.findAll('ember-flexberry-dummy-suggestion').then(posts => posts.toArray());
+  }
+});