+++ tests/dummy/app/routes/ember-flexberry-dummy-suggestion-edit.js
@@ -34,7 +34,7 @@ export default EditFormRoute.extend(EditFormRouteOperationsIndicationMixin, {
   developerUserSettings: {
     suggestionUserVotesGroupEdit: {
       'DEFAULT': {
-        'columnWidths': [{ 'propName': 'OlvRowToolbar', 'fixed': true }, { 'propName': 'voteType' },
+        'columnWidths': [{ 'propName': 'OlvRowToolbar', 'fixed': true, 'width': 82 }, { 'propName': 'voteType', 'width': 133 },
 		{ 'propName': 'author', 'width': 348 }, { 'propName': 'author.eMail', 'width': 531 }],
         'sorting': [{ 'propName': 'author', 'direction': 'asc' }]
       }